initial database design
Refactored `User` schema to store a `first_name` and `last_name` instead of a `username`. For a better UX, these fields have been made nullable so that the user can register by simply providing an email address and password. The app could then allow the user to set a first name and last name in a "edit profile" section.

Schemas defined for the the following models: `Project`, `ProjectInvite`, `ProjectMember`, `Task`.